feat: splited mcp server
